refactor(desktop): extract subscription model fetch
